Fix four portability defects that only a non-Windows CI could find
CI ran for the first time and failed in every job. All four causes are real, and
all four were invisible locally because clang-cl on Windows is the only compiler
this has ever been built with.

**Misaligned loads, reported by UBSan.** A raw-mode brick payload is a one-byte
mode tag followed by verbatim voxels, so the decoder reads it at
`payload.data() + 1`. For any dtype wider than a byte that base is misaligned by
construction -- on a u16 volume it is always odd -- and load_voxel cast it
straight to a `const std::uint16_t*`. x86 does not fault and MSVC never
complained, so it stood. It is still undefined behaviour: a strict-alignment
target faults on it, and both clang and gcc may assume the pointer is aligned
and vectorize on that basis, which is the kind of miscompile that shows up at
one optimization level and not another. load_voxel and store_voxel now go
through memcpy, which is defined for any alignment and which every compiler here
folds back into a single unaligned access.

**gcc built the whole project as C++17.** CMAKE_CXX_STANDARD was 26 with
STANDARD_REQUIRED OFF and a comment claiming that would "fall back to 23 on
older toolchains". It does not do that. When the compiler cannot provide the
requested standard, CMake with REQUIRED OFF omits the -std flag entirely and you
get the compiler default -- gnu++17 on gcc-13. clang and MSVC default to
something recent enough that nobody noticed. gcc-13 failed on a defaulted
operator== and on std::span not existing. It is C++23 and required now, which is
what the source actually targets: dct16.hpp explicitly avoids constexpr <cmath>
to stay buildable at 23. A toolchain that cannot manage it now fails at
configure time with a clear message.

**-Wconversion under -Werror.** Two subtractions in test_cuda.cpp cast only the
left operand to double and let the right convert implicitly. clang on Linux
diagnoses it, clang-cl did not.

**A shell quoting bug in the workflow I had just written.** The architecture list
is semicolon-separated and was interpolated unquoted, so bash split it and tried
to run `89-real` as a command.

Everything green locally on all four configurations after these -- sse2, avx2,
-Werror and CUDA -- but the point of this commit is the ones that cannot be
checked here, so the run is what decides.

+++ tests/test_cuda.cpp
@@ -235,7 +235,8 @@ TEST(cuda_rdo_matches_cpu_rdo) {
     REQUIRE(encode(raw.data(), v.dims, DType::u8, opts, cpu, Backend::cpu_scalar) == Status::ok);
     REQUIRE(encode(raw.data(), v.dims, DType::u8, opts, gpu, Backend::cuda) == Status::ok);
 
-    const double dsize = std::fabs(static_cast<double>(cpu.size()) - gpu.size()) /
+    const double dsize = std::fabs(static_cast<double>(cpu.size()) -
+                                   static_cast<double>(gpu.size())) /
                          static_cast<double>(cpu.size());
     std::printf("       q=%.2f  cpu %zu B  gpu %zu B  (%.4f%% apart)\n",
                 static_cast<double>(q), cpu.size(), gpu.size(), dsize * 100.0);
@@ -293,7 +294,8 @@ TEST(cuda_encode_matches_cpu) {
     REQUIRE(encode(raw.data(), v.dims, DType::u8, opts, cpu_r, Backend::cpu_scalar) ==
             Status::ok);
     REQUIRE(encode(raw.data(), v.dims, DType::u8, opts, gpu_r, Backend::cuda) == Status::ok);
-    const double dsize = std::fabs(static_cast<double>(cpu_r.size()) - gpu_r.size()) /
+    const double dsize = std::fabs(static_cast<double>(cpu_r.size()) -
+                                   static_cast<double>(gpu_r.size())) /
                          static_cast<double>(cpu_r.size());
     std::printf("       q=%.2f rdo: cpu %zu B, gpu %zu B (%.4f%% apart)\n",
                 static_cast<double>(q), cpu_r.size(), gpu_r.size(), dsize * 100.0);
